requirements.txt
pyautogui
Pillow
pytesseract
opencv-python-headless
numpy
pygetwindow
pyperclip
keyboard
tesseract